# swagger.yaml
openapi: 3.0.0
info:
  title: Image Processing API
  description: Proimage image proceessing creation, retieaval and formatting of images [CommonMark](http://commonmark.org/help/) or HTML.
  version: 0.1.9
servers:
  - url: https://proimage-akups.herokuapp.com
    description: Optional server description, e.g. Main (production) server
paths:
  /image:
    post:
      summary: Saves an image.
      description: Creates an image file on S3 and creates a record in the database.
      responses:
        "200": # status code
          description: A JSON object with the details of the created image
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    example: 5fd0bfb68c6e8e5df2562e3c
                  id:
                    type: string
                    example: c2092c38-deb8-48d1-a9c6-d7bc7b2fa73a
                  created_at:
                    type: string
                    example: 2020-12-09T12:14:46.915Z
                  name:
                    type: string
                    example: tweety.jpg
  /image/:imageId:
    get:
      summary: Retrieves saved image.
      description: Retrieves an image a record in the database with a specific id.
      parameters:
        - name: imageId
          in: path
          required: true
          schema:
            type: string
            example: c2092c38-deb8-48d1-a9c6-d7bc7b2fa73a
          description: The id of the image to be retrieved.
      responses:
        "200": # status code
          description: A JSON object with the details of the retrieved image
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    example: 5fd0bfb68c6e8e5df2562e3c
                  id:
                    type: string
                    example: c2092c38-deb8-48d1-a9c6-d7bc7b2fa73a
                  created_at:
                    type: string
                    example: 2020-12-09T12:14:46.915Z
                  name:
                    type: string
                    example: tweety.jpg
                  s3Key:
                    type: string
                    example: c2092c38-deb8-48d1-a9c6-d7bc7b2fa73a_tweety.jpg
                  url:
                    type: string
                    example: https://classic-cartoons.s3.eu-central-1.amazonaws.com/c2092c38-deb8-48d1-a9c6-d7bc7b2fa73a_tweety.jpg